Fix Arc issue with Kotlin Serialization in RESTEasy Reactive and REST Client
Fixes: #22712
