+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-kotlin-serialization/deployment/src/main/java/io/quarkus/kotlin/serialization/deployment/KotlinSerializationProcessor.java
@@ -1,28 +1,21 @@
 package io.quarkus.kotlin.serialization.deployment;
 
 import static io.quarkus.deployment.Feature.RESTEASY_REACTIVE_KOTLIN_SERIALIZATION;
-import static io.quarkus.deployment.annotations.ExecutionTime.STATIC_INIT;
 import static io.quarkus.resteasy.reactive.common.deployment.ServerDefaultProducesHandlerBuildItem.json;
 
 import java.util.List;
 
-import javax.inject.Singleton;
 import javax.ws.rs.core.MediaType;
 
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.arc.deployment.SyntheticBeanBuildItem;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
-import io.quarkus.kotlin.serialization.KotlinSerializationConfig;
 import io.quarkus.kotlin.serialization.KotlinSerializationMessageBodyReader;
 import io.quarkus.kotlin.serialization.KotlinSerializationMessageBodyWriter;
-import io.quarkus.kotlin.serialization.KotlinSerializerRecorder;
 import io.quarkus.resteasy.reactive.common.deployment.ServerDefaultProducesHandlerBuildItem;
 import io.quarkus.resteasy.reactive.spi.MessageBodyReaderBuildItem;
 import io.quarkus.resteasy.reactive.spi.MessageBodyWriterBuildItem;
-import kotlinx.serialization.json.Json;
 
 public class KotlinSerializationProcessor {
     @BuildStep
@@ -40,16 +33,6 @@ public class KotlinSerializationProcessor {
         additionalWriters.produce(new MessageBodyWriterBuildItem(
                 KotlinSerializationMessageBodyWriter.class.getName(), Object.class.getName(), List.of(
                         MediaType.APPLICATION_JSON)));
-    }
-
-    @BuildStep
-    @Record(STATIC_INIT)
-    public SyntheticBeanBuildItem createJson(KotlinSerializerRecorder recorder, KotlinSerializationConfig config) {
-        return SyntheticBeanBuildItem
-                .configure(Json.class)
-                .scope(Singleton.class)
-                .supplier(recorder.configFactory(config))
-                .unremovable().done();
     }
 
     @BuildStep
